+++ src/client/src/pages/detail/form/stratigraphy/lithology/lithologyLayers/descriptionLayers/descriptionLayers.jsx
@@ -66,7 +66,7 @@ const DescriptionLayers = props => {
     // include empty items in description column to signal missing descriptions
     const tempDescriptions = [];
     descriptions
-      .sort((a, b) => a.fromDepth - b.fromDepth)
+      ?.sort((a, b) => a.fromDepth - b.fromDepth)
       .forEach((description, index) => {
         const expectedFromDepth = index === 0 ? 0 : descriptions[index - 1]?.toDepth;
         if (description.fromDepth !== expectedFromDepth) {